refactor(components): extract form primitives + collapse barrel (Phase 6 #5)
Splits Form, FormField, TextInput, and DatePicker out of builtins.tsx
into src/components/primitives/form.tsx (carrying the formValues
helper alongside Form, where it's used).

builtins.tsx is now a 9-line re-export barrel that pulls all 19
primitives from src/components/primitives/. Existing call sites
(A2UIRenderer.tsx, builtins.test.tsx) keep importing from
./builtins with no source changes; tests pass unchanged.

This completes the file split planned in #43. No logic changes.
